@@ -1,72 +1,68 @@
 {
   "name": "admin",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "postinstall": "ngcc --properties es2015 browser module main --first-only",
     "start": "ng serve",
-    "build": "ng build --prod --aot --build-optimizer --common-chunk --vendor-chunk --optimization --progress",
+    "build": "ng build --configuration production",
+    "build:prod": "ng build --configuration production --base-href='/admin/'",
+    "lite-server": "lite-server --baseDir='dist'",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.1.0",
-    "@angular/common": "^11.1.0",
-    "@angular/compiler": "^11.1.0",
-    "@angular/core": "^11.1.0",
-    "@angular/forms": "^11.2.0",
-    "@angular/platform-browser": "^11.1.0",
-    "@angular/platform-browser-dynamic": "^11.1.0",
-    "@angular/router": "^11.1.0",
+    "@angular/animations": "~12.0.4",
+    "@angular/common": "~12.0.4",
+    "@angular/compiler": "~12.0.4",
+    "@angular/core": "~12.0.4",
+    "@angular/forms": "~12.0.4",
+    "@angular/platform-browser": "~12.0.4",
+    "@angular/platform-browser-dynamic": "~12.0.4",
+    "@angular/router": "~12.0.4",
+    "@aws-amplify/ui-angular": "^1.0.9",
     "@coreui/angular": "2.9.4",
     "@coreui/coreui": "2.1.16",
     "@coreui/coreui-plugin-chartjs-custom-tooltips": "1.3.1",
     "@coreui/icons": "0.3.0",
-    "angular-auth-oidc-client": "^11.2.4",
+    "aws-amplify": "^4.0.2",
     "bootstrap": "^4.5.0",
     "chart.js": "^2.9.3",
     "core-js": "^3.6.5",
     "flag-icon-css": "^3.5.0",
     "font-awesome": "^4.7.0",
+    "is-svg": "^4.2.2",
+    "lite-server": "^2.6.1",
     "moment": "^2.27.0",
     "mutationobserver-shim": "^0.3.7",
     "ng2-charts": "^2.3.2",
-    "ngx-bootstrap": "^5.6.1",
-    "ngx-perfect-scrollbar": "^9.0.0",
-    "rxjs": "^6.6.0",
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-perfect-scrollbar": "^10.1.1",
+    "rxjs": "~6.6.0",
     "simple-line-icons": "^2.5.2",
-    "ts-helpers": "^1.1.2",
+    "ssri": "^8.0.1",
     "tslib": "^2.0.0",
     "web-animations-js": "^2.3.2",
-    "zone.js": "^0.11.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1101.0",
-    "@angular/cli": "^11.1.0",
-    "@angular/compiler-cli": "^11.1.0",
-    "@angular/language-service": "^11.1.0",
-    "@types/jasmine": "^3.6.0",
-    "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^13.13.14",
-    "browserslist": "^4.16.6",
+    "@angular-devkit/build-angular": "~12.0.4",
+    "@angular/cli": "~12.0.4",
+    "@angular/compiler-cli": "~12.0.4",
+    "@types/jasmine": "~3.6.0",
+    "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "dns-packet": "^1.3.4",
-    "is-svg": "^4.2.2",
-    "jasmine-core": "^3.6.0",
-    "jasmine-spec-reporter": "^5.0.2",
-    "karma": "^6.3.2",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^3.0.3",
-    "karma-jasmine": "^4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.4",
-    "protractor": "^7.0.0",
-    "ts-node": "^8.10.2",
-    "tslint": "^6.1.3",
-    "typescript": "~4.1.2"
-  },
-  "engines": {
-    "node": ">= 10.13",
-    "npm": ">= 6"
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.3",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.2.4"
   }
 }
