@@ -20,16 +20,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'serverless-saas-ref-arch-admin'`, () => {
+  it(`should have as title 'admin'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('serverless-saas-ref-arch-admin');
+    expect(app.title).toEqual('admin');
   });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.nativeElement;
-    expect(compiled.querySelector('.content span').textContent).toContain('serverless-saas-ref-arch-admin app is running!');
+    expect(compiled.querySelector('.content span').textContent).toContain('admin app is running!');
   });
 });
