@@ -1,30 +1,17 @@
 /*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * SPDX-License-Identifier: MIT-0
  */
+
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { environment } from '../../environments/environment';
-import {TenantService} from './tenant.service'
-
+import { Router } from '@angular/router';
+import { TenantService } from './tenant.service';
 
 @Component({
   selector: 'app-create',
   templateUrl: './create.component.html',
-  styles: [
-  ]
+  styles: [],
 })
 export class CreateComponent implements OnInit {
   form: FormGroup;
@@ -32,39 +19,36 @@ export class CreateComponent implements OnInit {
   error = false;
   success = false;
 
-  constructor(private fb: FormBuilder,
-              private tenantSvc: TenantService) { }
+  constructor(
+    private fb: FormBuilder,
+    private tenantSvc: TenantService,
+    private router: Router
+  ) {}
 
   ngOnInit(): void {
     this.form = this.fb.group({
       tenantName: [null, [Validators.required]],
       tenantAddress: [null],
       tenantEmail: [null, [Validators.email, Validators.required]],
-      tenantPhone: [null,[Validators.required]],
+      tenantPhone: [null, [Validators.required]],
       tenantTier: [null, [Validators.required]],
-
-      });
+    });
   }
 
   onSubmit() {
-
-    //this.form.value.tenantName = this.form.value.tenantName.replace(/\s/g, "").toLowerCase();
-
-
+    this.submitting = true;
     const user = {
-      ...this.form.value
+      ...this.form.value,
     };
 
-
-    this.submitting = true;
-    this.tenantSvc.createTenant(user)
-    .subscribe(
-      val => {
+    this.tenantSvc.createTenant(user).subscribe(
+      (val) => {
         this.submitting = false;
         this.success = true;
         this.error = false;
+        this.router.navigate(['tenants']);
       },
-      err => {
+      (err) => {
         this.submitting = false;
         this.success = false;
         this.error = true;
@@ -92,6 +76,4 @@ export class CreateComponent implements OnInit {
     const formField = this.form.get(field);
     return !!formField.errors[error];
   }
-
-
 }
