@@ -1,19 +1,8 @@
 /*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * SPDX-License-Identifier: MIT-0
  */
+
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { CognitoGuard } from '../cognito.guard';
@@ -21,43 +10,43 @@ import { CreateComponent } from './create.component';
 import { TenantDetailComponent } from './tenant-detail.component';
 import { TenantListComponent } from './tenant-list.component';
 
-
 const routes: Routes = [
-{
-  path: '',
-  data: {
-    title: 'Tenant Management'
+  {
+    path: '',
+    data: {
+      title: 'Tenant Management',
+    },
+    children: [
+      {
+        path: '',
+        component: TenantListComponent,
+        data: {
+          title: 'Tenant List',
+        },
+        canActivate: [CognitoGuard],
+      },
+      {
+        path: 'create',
+        component: CreateComponent,
+        data: {
+          title: 'Create New Tenant',
+        },
+        canActivate: [CognitoGuard],
+      },
+      {
+        path: ':id',
+        component: TenantDetailComponent,
+        data: {
+          title: 'Tenant Detail',
+        },
+        canActivate: [CognitoGuard],
+      },
+    ],
   },
-  children: [
-    {
-      path: '',
-      component: TenantListComponent,
-      data: {
-        title: 'Tenant List'
-      },
-      canActivate: [CognitoGuard],
-    },
-    {
-      path: 'create',
-      component: CreateComponent,
-      data: {
-        title: 'Create New Tenant'
-      },
-      canActivate: [CognitoGuard],
-    },
-    {
-      path: ':id',
-      component: TenantDetailComponent,
-      data: {
-        title: 'Tenant Detail'
-      },
-      canActivate: [CognitoGuard],
-    },
-  ]
-}];
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class TenantRoutingModule { }
+export class TenantRoutingModule {}
