@@ -1,42 +1,28 @@
 /*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * SPDX-License-Identifier: MIT-0
  */
-
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
+import { Tenant } from './models/tenant';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class TenantService {
-  constructor(private http: HttpClient) { }
-
-  apiUrl = `${environment.apiGatewayUrl}`;
-
-
+  constructor(private http: HttpClient) {}
+  baseUrl = `${environment.apiUrl}`;
+  tenantsApiUrl = `${this.baseUrl}/tenants`;
+  registrationApiUrl = `${this.baseUrl}/registration`;
 
   // TODO strongly-type these anys as tenants once we dial in what the tenant call should return
   getTenants(): Observable<any[]> {
-    return this.http.get<any[]>(this.apiUrl+'/tenants');
+    return this.http.get<any[]>(this.tenantsApiUrl);
   }
 
-
-  createTenant(user: any): Observable<any> {
-    return this.http.post<any>(this.apiUrl+'/registration', user);
+  createTenant(tenant: any): Observable<any[]> {
+    return this.http.post<any[]>(this.registrationApiUrl, tenant);
   }
 }
