@@ -1,20 +1,7 @@
 /*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * SPDX-License-Identifier: MIT-0
  */
-
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { UsersService } from '../users.service';
@@ -22,23 +9,21 @@ import { UsersService } from '../users.service';
 @Component({
   selector: 'app-user-create',
   templateUrl: './user-create.component.html',
-  styles: [
-  ]
+  styles: [],
 })
 export class UserCreateComponent implements OnInit {
   userForm: FormGroup;
   error: boolean = false;
   success: boolean = false;
 
-  constructor(private fb: FormBuilder, private userSvc: UsersService) {
-  }
+  constructor(private fb: FormBuilder, private userSvc: UsersService) {}
 
   ngOnInit(): void {
     this.userForm = this.fb.group({
       userName: [null, [Validators.required]],
       userEmail: [null, [Validators.email, Validators.required]],
       tenantId: [null, [Validators.required]],
-      userRole: [null,[Validators.required]]
+      userRole: [null, [Validators.required]],
     });
   }
 
@@ -65,15 +50,14 @@ export class UserCreateComponent implements OnInit {
   onSubmit() {
     this.userForm.value.tenantId = this.userForm.value.tenantId.trim();
     const user = {
-      ...this.userForm.value
+      ...this.userForm.value,
     };
-    this.userSvc.create(user)
-    .subscribe(
-      val => {
+    this.userSvc.create(user).subscribe(
+      (val) => {
         this.success = true;
         this.error = false;
       },
-      err => {
+      (err) => {
         this.success = false;
         this.error = true;
         console.log(err);
