@@ -1,18 +1,6 @@
 /*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * SPDX-License-Identifier: MIT-0
  */
 import { Component, OnInit } from '@angular/core';
 import { getStyle, hexToRgba } from '@coreui/coreui/dist/js/coreui-utilities';
@@ -30,7 +18,7 @@ interface ChartData {
 }
 
 @Component({
-  templateUrl: 'dashboard.component.html'
+  templateUrl: 'dashboard.component.html',
 })
 export class DashboardComponent implements OnInit {
   constructor(private tenantSvc: TenantService) {}
@@ -40,41 +28,52 @@ export class DashboardComponent implements OnInit {
   public lineChart2Data: Array<any> = [
     {
       data: [1, 18, 9, 17, 34, 22, 11],
-      label: 'Series A'
-    }
+      label: 'Series A',
+    },
   ];
-  public lineChart2Labels: Array<any> = ['January', 'February', 'March', 'April', 'May', 'June', 'July'];
+  public lineChart2Labels: Array<any> = [
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+  ];
   public lineChart2Options: any = {
     tooltips: {
       enabled: false,
-      custom: CustomTooltips
+      custom: CustomTooltips,
     },
     maintainAspectRatio: false,
     scales: {
-      xAxes: [{
-        gridLines: {
-          color: 'transparent',
-          zeroLineColor: 'transparent'
+      xAxes: [
+        {
+          gridLines: {
+            color: 'transparent',
+            zeroLineColor: 'transparent',
+          },
+          ticks: {
+            fontSize: 2,
+            fontColor: 'transparent',
+          },
         },
-        ticks: {
-          fontSize: 2,
-          fontColor: 'transparent',
-        }
-
-      }],
-      yAxes: [{
-        display: false,
-        ticks: {
+      ],
+      yAxes: [
+        {
           display: false,
-          min: 1-5,
-          max: 50+5,
-        }
-      }],
+          ticks: {
+            display: false,
+            min: 1 - 5,
+            max: 50 + 5,
+          },
+        },
+      ],
     },
     elements: {
       line: {
         tension: 0.00001,
-        borderWidth: 1
+        borderWidth: 1,
       },
       point: {
         radius: 4,
@@ -83,14 +82,15 @@ export class DashboardComponent implements OnInit {
       },
     },
     legend: {
-      display: false
-    }
+      display: false,
+    },
   };
   public lineChart2Colours: Array<any> = [
-    { // grey
+    {
+      // grey
       backgroundColor: getStyle('--info'),
-      borderColor: 'rgba(255,255,255,.55)'
-    }
+      borderColor: 'rgba(255,255,255,.55)',
+    },
   ];
   public lineChart2Legend = false;
   public lineChart2Type = 'line';
@@ -104,19 +104,48 @@ export class DashboardComponent implements OnInit {
   public mainChartData: Array<any> = [
     {
       data: this.mainChartData1,
-      label: 'Current'
+      label: 'Current',
     },
     {
       data: this.mainChartData2,
-      label: 'Previous'
+      label: 'Previous',
     },
     {
       data: this.mainChartData3,
-      label: 'BEP'
-    }
+      label: 'BEP',
+    },
   ];
   /* tslint:disable:max-line-length */
-  public mainChartLabels: Array<any> = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday', 'Monday', 'Thursday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'];
+  public mainChartLabels: Array<any> = [
+    'Monday',
+    'Tuesday',
+    'Wednesday',
+    'Thursday',
+    'Friday',
+    'Saturday',
+    'Sunday',
+    'Monday',
+    'Tuesday',
+    'Wednesday',
+    'Thursday',
+    'Friday',
+    'Saturday',
+    'Sunday',
+    'Monday',
+    'Tuesday',
+    'Wednesday',
+    'Thursday',
+    'Friday',
+    'Saturday',
+    'Sunday',
+    'Monday',
+    'Thursday',
+    'Wednesday',
+    'Thursday',
+    'Friday',
+    'Saturday',
+    'Sunday',
+  ];
   /* tslint:enable:max-line-length */
   public mainChartOptions: any = {
     tooltips: {
@@ -126,70 +155,79 @@ export class DashboardComponent implements OnInit {
       mode: 'index',
       position: 'nearest',
       callbacks: {
-        labelColor: function(tooltipItem, chart) {
-          return { backgroundColor: chart.data.datasets[tooltipItem.datasetIndex].borderColor };
-        }
-      }
+        labelColor: function (tooltipItem, chart) {
+          return {
+            backgroundColor:
+              chart.data.datasets[tooltipItem.datasetIndex].borderColor,
+          };
+        },
+      },
     },
     responsive: true,
     maintainAspectRatio: false,
     scales: {
-      xAxes: [{
-        gridLines: {
-          drawOnChartArea: false,
+      xAxes: [
+        {
+          gridLines: {
+            drawOnChartArea: false,
+          },
+          ticks: {
+            callback: function (value: any) {
+              return value.charAt(0);
+            },
+          },
         },
-        ticks: {
-          callback: function(value: any) {
-            return value.charAt(0);
-          }
-        }
-      }],
-      yAxes: [{
-        ticks: {
-          beginAtZero: true,
-          maxTicksLimit: 5,
-          stepSize: Math.ceil(250 / 5),
-          max: 250
-        }
-      }]
+      ],
+      yAxes: [
+        {
+          ticks: {
+            beginAtZero: true,
+            maxTicksLimit: 5,
+            stepSize: Math.ceil(250 / 5),
+            max: 250,
+          },
+        },
+      ],
     },
     elements: {
       line: {
-        borderWidth: 2
+        borderWidth: 2,
       },
       point: {
         radius: 0,
         hitRadius: 10,
         hoverRadius: 4,
         hoverBorderWidth: 3,
-      }
+      },
     },
     legend: {
-      display: false
-    }
+      display: false,
+    },
   };
   public mainChartColours: Array<any> = [
-    { // brandInfo
-      backgroundColor: hexToRgba(getStyle('--info'), 10),
-      borderColor: getStyle('--info'),
-      pointHoverBackgroundColor: '#fff'
+    {
+      // brandInfo
+      backgroundColor: '#20a8d8',
+      borderColor: '#20a8d8',
+      pointHoverBackgroundColor: '#fff',
     },
-    { // brandSuccess
+    {
+      // brandSuccess
       backgroundColor: 'transparent',
       borderColor: getStyle('--success'),
-      pointHoverBackgroundColor: '#fff'
+      pointHoverBackgroundColor: '#fff',
     },
-    { // brandDanger
+    {
+      // brandDanger
       backgroundColor: 'transparent',
       borderColor: getStyle('--danger'),
       pointHoverBackgroundColor: '#fff',
       borderWidth: 1,
-      borderDash: [8, 5]
-    }
+      borderDash: [8, 5],
+    },
   ];
   public mainChartLegend = false;
   public mainChartType = 'line';
-
 
   public random(min: number, max: number) {
     return Math.floor(Math.random() * (max - min + 1) + min);
@@ -197,25 +235,24 @@ export class DashboardComponent implements OnInit {
 
   ngOnInit(): void {
     // generate random values for mainChart
-    // this.tenantSvc.getTenants()
-    // .subscribe(val => {
-    //   val.forEach(tenant => {
-    //     let simulatedOrders = [];
-    //     //simulate 7 months of orders
-    //     for(let j = 0; j < 7; j++) {
-    //       simulatedOrders.push(this.random(0, 50));
-    //     }
-    //     const dataSet: DataSet = {
-    //       label: 'Monthly Orders',
-    //       data: simulatedOrders,
-    //     }
-    //     this.data.push({
-    //       tenantId: tenant.tenantId,
-    //       dataSet: [dataSet],
-    //       totalOrders: simulatedOrders.reduce((a,b) => a+b, 0)
-    //     });
-    //   });
-    // });
+    this.tenantSvc.getTenants().subscribe((val) => {
+      val.forEach((tenant) => {
+        let simulatedOrders = [];
+        //simulate 7 months of orders
+        for (let j = 0; j < 7; j++) {
+          simulatedOrders.push(this.random(0, 50));
+        }
+        const dataSet: DataSet = {
+          label: 'Monthly Orders',
+          data: simulatedOrders,
+        };
+        this.data.push({
+          tenantId: tenant.tenantId,
+          dataSet: [dataSet],
+          totalOrders: simulatedOrders.reduce((a, b) => a + b, 0),
+        });
+      });
+    });
 
     for (let i = 0; i <= this.mainChartElements; i++) {
       this.mainChartData1.push(this.random(50, 200));
