@@ -1,7 +1,8 @@
 export const environment = {
   production: true,
-  clientId: 'ii333sisfk7iih9ioischa59i',
-  issuer: 'https://cognito-idp.us-west-2.amazonaws.com/us-west-2_DE4zpiGWm',
-  apiGatewayUrl: 'https://3l6yegpnyc.execute-api.us-west-2.amazonaws.com/prod/',
-  domain: ''
+  clientId: '598qmpppgi26dqkoef4p2rrh3g',
+  issuer: 'https://cognito-idp.us-west-2.amazonaws.com/us-west-2_eItw7wpUO',
+  apiGatewayUrl: 'https://wchv3gaofb.execute-api.us-west-2.amazonaws.com/prod',
+  apiUrl: 'https://wchv3gaofb.execute-api.us-west-2.amazonaws.com/prod',
+  domain: '',
 };
